add scenario tests framework
Utilizing pytest and jinja2 templating to simplify the process in
developing scenario tests. Test developers can easily create request
templates and load them into JSON request objects by provide only the
required fields. The templates, along with their corresponding fixtures
are designed to be reusable. With templating, the goal is to reduce the
complexity of having to handcraft the JSON request in every test.