updates to test_create_edit_network test
currently the test_create_edit_network updates the vlan id
but does not revert the change once the test is done.

This PR creates a new fixture at the class scope level
which creates a network with vlan_id +1 and the test
updates it with -1. Once the test is done, this network
is deleted.