network interact tests: fix SSH check
- Fix SSH connection checks
There are serveral places where a test checks that an SSH connection to
a VM on a specific interface can not be established.
Doing so by checking for the specific error "no route to host" is
potentially wrong, because this presumes that the host executing the
test even knows how to route to the network. This isn't always the case,
e.g. when executing the test suite on a developer workstation instead of
a CI server. In those cases there may be an SSH error "network
unreachable" instead, which should also be acceptable to let the test
pass.
To allow for this case, instead of checking the output of the SSH
command for a specific substring, checking for the return code 255 is
better, because it allows for different kind of errors and is less
fragile than searching for a string in the command output.

- Mark networking tests with @pytest.mark.networks
This includes the network interaction tests when selecting tests with
the pytest mark "networks"

- Fix SSH connection with paramiko when the ssh agent contains keys, but
password authentication is used
By defaul paramiko will try to use key based authentication using
whichever keys it finds in the ssh-agent. In this case however password
based authentication is used and paramiko should not try to use the keys
from the ssh-agent. Otherwise it may be confused and error out.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>