tests: test CRD versions
Add test which checks if the CRD versions defined in the cluster match
the versions in the subcharts in the git repo.
This test ensures that the subcharts in the release version don't
unexpectedly change their API versions within a minor Harvester release.

To implement this test, additional infrastructure for the Kubernetes API
has been added.

fixes: https://github.com/harvester/tests/issues/1314

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>